backend/requirements.txt
# backend/requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
h11==0.16.0
sqlalchemy==2.0.46
pymysql==1.1.2
python-dotenv==1.2.1
passlib==1.7.4
PyJWT==2.11.0

# Testing
pytest==8.0.0
pytest-asyncio==0.21.1
httpx==0.25.2
